merge

+++ scripts/clash_config.sh
@@ -37,32 +37,36 @@ prepare(){
 
 case $ACTION in
 start)
-	if [ "$dnsmasqplan" == "overwrite" ]; then
-		[ ! -L "/jffs/scripts/dnsmasq.postconf" ] && ln -sf /koolshare/merlinclash/conf/dnsmasq.postconf /jffs/scripts/dnsmasq.postconf && echo_date "创建dnsmasq.postconf软链接" >> $LOG_FILE
+	if [ "$mcenable" == "1" ];then
+		if [ "$dnsmasqplan" == "overwrite" ]; then
+			[ ! -L "/jffs/scripts/dnsmasq.postconf" ] && ln -sf /koolshare/merlinclash/conf/dnsmasq.postconf /jffs/scripts/dnsmasq.postconf && echo_date "创建dnsmasq.postconf软链接" >> $LOG_FILE
 
-		sh /koolshare/merlinclash/clashconfig.sh start >> /tmp/upload/merlinclash_log.txt
-	else
-		rm -rf /jffs/scripts/dnsmasq.postconf
-		prepare
-		sed -i '$a no-resolv' /etc/dnsmasq.conf
-		sed -i '$a servers-file=/tmp/resolv.dnsmasq' /etc/dnsmasq.conf
-		sh /koolshare/merlinclash/clashconfig_0101.sh start >> /tmp/upload/merlinclash_log.txt
+			sh /koolshare/merlinclash/clashconfig.sh start >> /tmp/upload/merlinclash_log.txt
+		else
+			rm -rf /jffs/scripts/dnsmasq.postconf
+			prepare
+			sed -i '$a no-resolv' /etc/dnsmasq.conf
+			sed -i '$a servers-file=/tmp/resolv.dnsmasq' /etc/dnsmasq.conf
+			sh /koolshare/merlinclash/clashconfig_0101.sh start >> /tmp/upload/merlinclash_log.txt
+		fi
 	fi
 	;;
 start_nat)
-	if [ "$dnsmasqplan" == "overwrite" ]; then
-		[ ! -L "/jffs/scripts/dnsmasq.postconf" ] && ln -sf /koolshare/merlinclash/conf/dnsmasq.postconf /jffs/scripts/dnsmasq.postconf && echo_date "创建dnsmasq.postconf软链接" >> $LOG_FILE
-		if [ "$dnsgoclash" == "1" ]; then
-			sh /koolshare/merlinclash/clashconfig.sh restart >> /tmp/upload/merlinclash_log.txt
+	if [ "$mcenable" == "1" ];then
+		if [ "$dnsmasqplan" == "overwrite" ]; then
+			[ ! -L "/jffs/scripts/dnsmasq.postconf" ] && ln -sf /koolshare/merlinclash/conf/dnsmasq.postconf /jffs/scripts/dnsmasq.postconf && echo_date "创建dnsmasq.postconf软链接" >> $LOG_FILE
+			if [ "$dnsgoclash" == "1" ]; then
+				sh /koolshare/merlinclash/clashconfig.sh restart >> /tmp/upload/merlinclash_log.txt
+			else
+				sh /koolshare/merlinclash/clashconfig.sh start_nat >> /tmp/upload/merlinclash_log.txt
+			fi
 		else
-			sh /koolshare/merlinclash/clashconfig.sh start_nat >> /tmp/upload/merlinclash_log.txt
+			rm -rf /jffs/scripts/dnsmasq.postconf
+			prepare
+			sed -i '$a no-resolv' /etc/dnsmasq.conf
+			sed -i '$a servers-file=/tmp/resolv.dnsmasq' /etc/dnsmasq.conf
+			sh /koolshare/merlinclash/clashconfig_0101.sh start_nat >> /tmp/upload/merlinclash_log.txt
 		fi
-	else
-		rm -rf /jffs/scripts/dnsmasq.postconf
-		prepare
-		sed -i '$a no-resolv' /etc/dnsmasq.conf
-		sed -i '$a servers-file=/tmp/resolv.dnsmasq' /etc/dnsmasq.conf
-		sh /koolshare/merlinclash/clashconfig_0101.sh start_nat >> /tmp/upload/merlinclash_log.txt
 	fi
 	;;
 esac
